optiq_flow/requirements.txt
qiskit>=1.0.0
qiskit-aer>=0.13.3
qiskit-ibm-runtime>=0.21.1
qiskit-algorithms>=0.3.0
qiskit-optimization>=0.6.0
yfinance>=0.2.36
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0
matplotlib>=3.8.0
plotly>=5.19.0
streamlit>=1.31.0
python-dotenv>=1.0.1
networkx>=3.2.1
cvxpy>=1.4.2
